django_com_mysql/settings.py
import os
from pathlib import Path
from decouple import config
import dj_database_url

DATABASES = {
    'default': dj_database_url.config()
}

# Caminho base do projeto
BASE_DIR = Path(__file__).resolve().parent.parent

# Segurança
SECRET_KEY = 'django-insecure-d@pq1f!7-3c=qna+#o0p31u_-ma2=*b&(a9wd1e7!b^b+tt46j'
DEBUG = False
ALLOWED_HOSTS = ['*']  # Em produção, substitua por seu domínio

# Aplicativos instalados
INSTALLED_APPS = [
    'django.contrib.admin',
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'django.contrib.messages',
    'django.contrib.staticfiles',

    # Apps do projeto
    'core',

    # Terceiros
    'django_bootstrap5',
    'versatileimagefield',
]

# Middlewares
MIDDLEWARE = [
    'django.middleware.security.SecurityMiddleware',
    'whitenoise.middleware.WhiteNoiseMiddleware',
    'django.contrib.sessions.middleware.SessionMiddleware',
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
    'django.contrib.messages.middleware.MessageMiddleware',
    'django.middleware.clickjacking.XFrameOptionsMiddleware',
]

# URLs
ROOT_URLCONF = 'django_com_mysql.urls'

# Templates
TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [BASE_DIR / 'core/templates'],  # diretório customizado
        'APP_DIRS': True,
        'OPTIONS': {
            'context_processors': [
                'django.template.context_processors.request',
                'django.contrib.auth.context_processors.auth',
                'django.contrib.messages.context_processors.messages',
            ],
        },
    },
]

# WSGI
WSGI_APPLICATION = 'django_com_mysql.wsgi.application'

# Banco de Dados (MySQL com variáveis de ambiente)

# Validação de senha
AUTH_PASSWORD_VALIDATORS = [
    {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator'},
    {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator'},
    {'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator'},
    {'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator'},
]

# Internacionalização
LANGUAGE_CODE = 'pt-br'
TIME_ZONE = 'America/Sao_Paulo'
USE_I18N = True
USE_TZ = True

# Arquivos estáticos
STATIC_URL = 'static/'
STATIC_ROOT = os.path.join(BASE_DIR, 'staticfiles')

# Arquivos de mídia (imagens enviadas)
MEDIA_URL = '/media/'
MEDIA_ROOT = os.path.join(BASE_DIR, 'midia')

# Campo padrão para chaves primárias
DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'

# Email (console no dev)
#EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'

# Configurações do VersatileImageField
VERSATILEIMAGEFIELD_SETTINGS = {
    'create_images_on_demand': True,
}
VERSATILEIMAGEFIELD_RENDITION_KEY_SETS = {
    'produto_image': [  # Nome do conjunto
        ('thumb', 'crop__124x124'),
        ('medium', 'crop__300x300'),
        ('large', 'crop__800x800'),
    ]
}
