requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-bootstrap5==25.1
django-versatileimagefield==3.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-magic==0.4.27
sqlparse==0.5.3
whitenoise==6.9.0
